refactor: pull up LogEventBuilder class
